requirements.txt
numpy>=1.23
imageio>=2.31
matplotlib>=3.7
pyyaml>=6.0
pytest>=7.2
mypy>=1.5